+++ src/main/java/net/ttddyy/dsproxy/proxy/JdbcProxyFactory.java
@@ -8,6 +8,7 @@ import javax.sql.DataSource;
 import java.sql.CallableStatement;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.Statement;
 
 /**
@@ -38,4 +39,26 @@ public interface JdbcProxyFactory {
     CallableStatement createCallableStatement(CallableStatement callableStatement, String query,
                                               InterceptorHolder interceptorHolder, ConnectionInfo connectionInfo,
                                               Connection proxyConnection);
+
+    /**
+     * Flag this factory instance to create a proxy for {@link ResultSet}.
+     *
+     * @param createResultSetProxy true to proxy ResultSet
+     * @return this instance
+     * @since 1.4.3
+     */
+    JdkJdbcProxyFactory createResultSetProxy(boolean createResultSetProxy);
+
+    /**
+     * Create a proxy for {@link ResultSet} if enabled.
+     *
+     * When {@link #createResultSetProxy(boolean)} is set to {@code true}, this method returns a proxied
+     * {@link ResultSet}. If {@code false} is set, passed original {@link ResultSet} is returned.
+     *
+     * @param resultSet a result set
+     * @return proxied ResultSet if enabled
+     * @since 1.4.3
+     */
+    ResultSet createResultSet(ResultSet resultSet);
+
 }