+++ src/main/java/net/ttddyy/dsproxy/proxy/jdk/CallableStatementInvocationHandler.java
@@ -2,6 +2,7 @@ package net.ttddyy.dsproxy.proxy.jdk;
 
 import net.ttddyy.dsproxy.ConnectionInfo;
 import net.ttddyy.dsproxy.proxy.InterceptorHolder;
+import net.ttddyy.dsproxy.proxy.JdbcProxyFactory;
 import net.ttddyy.dsproxy.proxy.PreparedStatementProxyLogic;
 
 import java.lang.reflect.InvocationHandler;
@@ -23,13 +24,14 @@ public class CallableStatementInvocationHandler implements InvocationHandler {
 
     public CallableStatementInvocationHandler(
             CallableStatement cs, String query, InterceptorHolder interceptorHolder, ConnectionInfo connectionInfo,
-            Connection proxyConnection) {
+            Connection proxyConnection, JdbcProxyFactory proxyFactory) {
         this.delegate = PreparedStatementProxyLogic.Builder.create()
                 .preparedStatement(cs)
                 .query(query)
                 .interceptorHolder(interceptorHolder)
                 .connectionInfo(connectionInfo)
                 .proxyConnection(proxyConnection)
+                .proxyFactory(proxyFactory)
                 .build();
     }