+++ src/main/java/net/ttddyy/dsproxy/proxy/jdk/StatementInvocationHandler.java
@@ -2,6 +2,7 @@ package net.ttddyy.dsproxy.proxy.jdk;
 
 import net.ttddyy.dsproxy.ConnectionInfo;
 import net.ttddyy.dsproxy.proxy.InterceptorHolder;
+import net.ttddyy.dsproxy.proxy.JdbcProxyFactory;
 import net.ttddyy.dsproxy.proxy.StatementProxyLogic;
 
 import java.lang.reflect.InvocationHandler;
@@ -19,12 +20,14 @@ public class StatementInvocationHandler implements InvocationHandler {
     private StatementProxyLogic delegate;
 
     public StatementInvocationHandler(
-            Statement stmt, InterceptorHolder interceptorHolder, ConnectionInfo connectionInfo, Connection proxyConnection) {
+            Statement stmt, InterceptorHolder interceptorHolder, ConnectionInfo connectionInfo,
+            Connection proxyConnection, JdbcProxyFactory proxyFactory) {
         this.delegate = StatementProxyLogic.Builder.create()
                 .statement(stmt)
                 .interceptorHolder(interceptorHolder)
                 .connectionInfo(connectionInfo)
                 .proxyConnection(proxyConnection)
+                .proxyFactory(proxyFactory)
                 .build();
     }