+++ src/test/java/net/ttddyy/dsproxy/proxy/StatementProxyLogicMockTest.java
@@ -3,17 +3,23 @@ package net.ttddyy.dsproxy.proxy;
 import net.ttddyy.dsproxy.ConnectionInfo;
 import net.ttddyy.dsproxy.ExecutionInfo;
 import net.ttddyy.dsproxy.QueryInfo;
+import net.ttddyy.dsproxy.listener.NoOpQueryExecutionListener;
 import net.ttddyy.dsproxy.listener.QueryExecutionListener;
+import net.ttddyy.dsproxy.proxy.jdk.JdkJdbcProxyFactory;
+import net.ttddyy.dsproxy.proxy.jdk.ResultSetInvocationHandler;
 import net.ttddyy.dsproxy.transform.QueryTransformer;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 
 import java.lang.reflect.Method;
+import java.lang.reflect.Proxy;
 import java.sql.Connection;
 import java.sql.ResultSet;
+import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.arrayContaining;
@@ -24,7 +30,9 @@ import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 import static org.hamcrest.Matchers.sameInstance;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -428,7 +436,12 @@ public class StatementProxyLogicMockTest {
         verifyListener(listener, "executeLargeUpdate", query, query);
     }
 
+
     private StatementProxyLogic getProxyLogic(Statement statement, QueryExecutionListener listener, Connection proxyConnection) {
+        return getProxyLogic(statement, listener, proxyConnection, JdbcProxyFactory.DEFAULT);
+    }
+
+    private StatementProxyLogic getProxyLogic(Statement statement, QueryExecutionListener listener, Connection proxyConnection, JdbcProxyFactory proxyFactory) {
         ConnectionInfo connectionInfo = new ConnectionInfo();
         connectionInfo.setDataSourceName(DS_NAME);
         InterceptorHolder interceptorHolder = new InterceptorHolder(listener, QueryTransformer.DEFAULT);
@@ -438,6 +451,7 @@ public class StatementProxyLogicMockTest {
                 .interceptorHolder(interceptorHolder)
                 .connectionInfo(connectionInfo)
                 .proxyConnection(proxyConnection)
+                .proxyFactory(proxyFactory)
                 .build();
     }
 
@@ -800,4 +814,64 @@ public class StatementProxyLogicMockTest {
         assertThat((Boolean) result, is(true));
     }
 
+    @Test
+    public void proxyResultSet() throws Throwable {
+
+        final AtomicReference<Object> listenerReceivedResult = new AtomicReference<Object>();
+        QueryExecutionListener listener = new NoOpQueryExecutionListener() {
+            @Override
+            public void afterQuery(ExecutionInfo execInfo, List<QueryInfo> queryInfoList) {
+                listenerReceivedResult.set(execInfo.getResult());
+            }
+        };
+
+
+        JdbcProxyFactory proxyFactory = new JdkJdbcProxyFactory();
+        proxyFactory.createResultSetProxy(true);
+
+        ResultSetMetaData metaData = mock(ResultSetMetaData.class);
+
+        ResultSet resultSet = mock(ResultSet.class);
+        when(resultSet.getMetaData()).thenReturn(metaData);
+
+
+        Statement stmt = mock(Statement.class);
+        when(stmt.executeQuery(anyString())).thenReturn(resultSet);
+        when(stmt.getGeneratedKeys()).thenReturn(resultSet);
+        when(stmt.getResultSet()).thenReturn(resultSet);
+        StatementProxyLogic logic = getProxyLogic(stmt, listener, null, proxyFactory);
+
+
+        // "executeQuery", "getGeneratedKeys", "getResultSet"
+        Method executeQueryMethod = Statement.class.getMethod("executeQuery", String.class);
+        Method getGeneratedKeysMethod = Statement.class.getMethod("getGeneratedKeys");
+        Method getResultSetMethod = Statement.class.getMethod("getResultSet");
+        Object result;
+
+        // check "executeQuery"
+        result = logic.invoke(executeQueryMethod, new Object[]{"SELECT *"});
+        assertThat(result, is(instanceOf(ResultSet.class)));
+        assertTrue(Proxy.isProxyClass(result.getClass()));
+        assertTrue(Proxy.getInvocationHandler(result).getClass().equals(ResultSetInvocationHandler.class));
+        assertThat("listener should receive proxied resultset", listenerReceivedResult.get(), sameInstance(result));
+
+        listenerReceivedResult.set(null);
+
+        // check "getGeneratedKeys"
+        result = logic.invoke(getGeneratedKeysMethod, null);
+        assertThat(result, is(instanceOf(ResultSet.class)));
+        assertTrue(Proxy.isProxyClass(result.getClass()));
+        assertTrue(Proxy.getInvocationHandler(result).getClass().equals(ResultSetInvocationHandler.class));
+        assertThat("listener should receive proxied resultset", listenerReceivedResult.get(), sameInstance(result));
+
+        listenerReceivedResult.set(null);
+
+        // check "getResultSet"
+        result = logic.invoke(getResultSetMethod, null);
+        assertThat(result, is(instanceOf(ResultSet.class)));
+        assertTrue(Proxy.isProxyClass(result.getClass()));
+        assertTrue(Proxy.getInvocationHandler(result).getClass().equals(ResultSetInvocationHandler.class));
+        assertThat("listener should receive proxied resultset", listenerReceivedResult.get(), sameInstance(result));
+
+    }
 }