+++ CHANGELOG.md
@@ -7,6 +7,16 @@
     e.g.: when `SLF4JQueryLoggingListener` writes SQL in DEBUG level, but the logger is set to INFO(more serious
           than DEBUG), then it will NOT perform logging logic including constructing log statement, etc.
 
+- Proxying ResultSet is refactored to align how other proxies are managed.
+  As part of refactoring, `ResultSetProxyJdbcProxyFactory` is removed.
+  To enable proxying ResultSet, use `ProxyDataSourceBuilder#proxyResultSet()`
+  ```java
+    // before
+    builder.jdbcProxyFactory(new ResultSetProxyJdbcProxyFactory()).build();
+    // new
+    builder.proxyResultSet().build();
+  ```
+
 
 ## 1.4.2